refactor: move search schema into `entities` lib
