+++ crates/entities/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod models;
+pub mod schema;
 pub mod test;
 
 pub use sea_orm;
